Rename index instead of dropping it (#1)
* [IMP] For small tables it's perfectly acceptable to drop index and then let Odoo recreate it, but with large tables this scenario is rather costly.

* [IMP] Dropping index in case new column is set to ``None``

+++ openupgradelib/openupgrade.py
@@ -507,13 +507,24 @@ def rename_columns(cr, column_spec):
     """
     for table in column_spec.keys():
         for (old, new) in column_spec[table]:
+            rename_index = True
+
             if new is None:
                 new = get_legacy_name(old)
+                rename_index = False
+
             logger.info("table %s, column %s: renaming to %s",
                         table, old, new)
             cr.execute(
                 'ALTER TABLE "%s" RENAME "%s" TO "%s"' % (table, old, new,))
-            cr.execute('DROP INDEX IF EXISTS "%s_%s_index"' % (table, old))
+
+            if rename_index:
+                cr.execute(
+                    'ALTER INDEX IF EXISTS "%s_%s_index" '
+                    'RENAME TO "%s_%s_index"' % (table, old, table, new)
+                )
+            else:
+                cr.execute('DROP INDEX IF EXISTS "%s_%s_index"' % (table, old))
 
 
 def rename_fields(env, field_spec, no_deep=False):
